feat: refactor ProductTable and useProductTable for improved search functionality and column management
