+++ src/components/product/ProductTable.jsx
@@ -1,59 +1,16 @@
 import {useProduct} from "../../hooks/product/useProduct.js";
-import {useEffect, useMemo} from "react";
-import {useProductTable} from "../../hooks/product/useProductTable.js";
-import {Box, Paper, TextField, Chip} from "@mui/material";
+import {useEffect} from "react";
+import {useProductTable} from "../../hooks/product/useProductTable.jsx";
+import {Box, Paper, TextField} from "@mui/material";
 import {DataGrid} from "@mui/x-data-grid";
 
 export const ProductTable = () => {
     const {products, handleGetProducts} = useProduct();
-    const {searchText, setSearchText, filteredProducts} = useProductTable(products);
+    const {searchText, setSearchText, filteredProducts, columns} = useProductTable(products);
 
     useEffect(() => {
         handleGetProducts();
-    }, []);
-
-    const columns = useMemo(() => [
-        {
-            field: "id",
-            headerName: "ID",
-            width: 90,
-        },
-        {
-            field: "name",
-            headerName: "Nombre",
-            width: 270,
-            sortable: true,
-        },
-        {
-            field: "category",
-            headerName: "Categoria",
-            width: 130,
-            sortable: true,
-        },
-        {
-            field: "price",
-            headerName: "Precio",
-            width: 100,
-        },
-        {
-            field: "unitMeasure",
-            headerName: "Unidad de medida",
-            width: 150,
-        },
-        {
-            field: "active",
-            headerName: "Activo",
-            width: 100,
-            renderCell: (params) => (
-                <Chip
-                    label={params.row.active ? "Si" : "No"}
-                    color={params.row.active ? "success" : "default"}
-                    variant="outlined"
-                    size="small"
-                />
-            ),
-        }
-    ], []);
+    }, [handleGetProducts]);
 
     return (
         <Paper sx={{p: 2, height: "100%"}}>
@@ -98,4 +55,4 @@ export const ProductTable = () => {
             </Box>
         </Paper>
     );
-}
+};
